requirements.txt
pyTelegramBotAPI
python-telegram-bot
schedule
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
